test setup for multi screen
