+++ code/ui.datagrid/src/datagrid.ts
@@ -1,4 +1,5 @@
 export * from './types';
 export * from './api/Editor/EditorContext';
 export * from './components/DataGrid';
+export * from './components/CellEditor';
 export { Grid, Cell } from './api';
