+++ code/ui.datagrid/test/cli/index.ts
@@ -1,13 +1,15 @@
+import { Subject } from 'rxjs';
+
 import { CommandState, t } from '../common';
 import { root } from './cmds';
 
-export function init(args: {}) {
-  const state = CommandState.create({
+export function init(args: { state$: Subject<Partial<t.ITestState>> }) {
+  const { state$ } = args;
+  return CommandState.create({
     root,
     getInvokeArgs: async state => {
-      return { props: { foo: 123 }, timeout: 5000 };
+      const props: t.ITestCommandProps = { state$ };
+      return { props };
     },
   });
-
-  return state;
 }
