+++ code/ui.datagrid/test/common/index.ts
@@ -1,6 +1,6 @@
 import '../../node_modules/@platform/css/reset.css';
 
-import * as t from './types';
+import * as t from '../types';
 export * from '../../src/common';
 export * from './libs';
