--- code/ui.datagrid/test/common/types.ts
@@ -1,4 +1,0 @@
-export * from '../../src/types';
-export * from '@platform/ui.cli/lib/types';
-
-export type ITestCommandProps = {};
