+++ code/ui.datagrid/test/components/Test.Grid.tsx
@@ -1,30 +1,22 @@
 import * as React from 'react';
 import { Subject } from 'rxjs';
-import { filter, map, takeUntil } from 'rxjs/operators';
+import { takeUntil } from 'rxjs/operators';
 
-import { datagrid, GlamorValue, Handsontable as HandsontableLib, t } from '../common';
-import { DebugEditor } from './Debug.Editor';
+import { Button, color, css, GlamorValue, ObjectView } from '../common';
+import { TestGridView } from './Test.Grid.view';
 
-export type ITestProps = {
-  style?: GlamorValue;
-  Table?: Handsontable;
-};
-export type ITestState = { values?: t.IGridValues };
-
-const DEFAULT = {
-  A1: 'A1',
-  B1: 'locked',
-  B2: 'cancel',
+export type ITestGridProps = { style?: GlamorValue };
+export type ITestGridState = {
+  data?: any;
 };
 
-export class Test extends React.PureComponent<ITestProps, ITestState> {
-  public state: ITestState = { values: DEFAULT };
-  public state$ = new Subject<Partial<ITestState>>();
+export class TestGrid extends React.PureComponent<ITestGridProps, ITestGridState> {
+  public state: ITestGridState = {};
   private unmounted$ = new Subject();
-  private events$ = new Subject<t.GridEvent>();
+  private state$ = new Subject<Partial<ITestGridState>>();
 
-  public datagrid!: datagrid.DataGrid;
-  private datagridRef = (ref: datagrid.DataGrid) => (this.datagrid = ref);
+  private testGrid!: TestGridView;
+  private testGridRef = (ref: TestGridView) => (this.testGrid = ref);
 
   /**
    * [Lifecycle]
@@ -34,60 +26,9 @@ export class Test extends React.PureComponent<ITestProps, ITestState> {
   }
 
   public componentDidMount() {
-    const events$ = this.events$.pipe(takeUntil(this.unmounted$));
-    // const keys$ = this.grid.keys$;
-
-    events$.pipe(filter(e => !['GRID/keydown'].includes(e.type))).subscribe(e => {
-      console.log('🌳 EVENT', e.type, e.payload);
-    });
-
-    events$
-      .pipe(
-        filter(e => e.type === 'GRID/EDITOR/end'),
-        map(e => e as t.IEndEditingEvent),
-        filter(e => !e.payload.isCancelled),
-      )
-      .subscribe(e => {
-        // console.log('HANDLE END  🐷 ');
-        // e.payload.cancel();
-      });
-
-    const beginEdit$ = events$.pipe(
-      filter(e => e.type === 'GRID/EDITOR/begin'),
-      map(e => e.payload as t.IBeginEditing),
-    );
-
-    beginEdit$.pipe(filter(e => e.cell.key === 'B1')).subscribe(e => {
-      // Cancel B1 edit operations before they begin.
-      e.cancel();
-    });
-
-    const change$ = events$.pipe(
-      filter(e => e.type === 'GRID/change'),
-      map(e => e.payload as t.IGridChange),
-    );
-
-    const changeSet$ = events$.pipe(
-      filter(e => e.type === 'GRID/changeSet'),
-      map(e => e.payload as t.IGridChangeSet),
-    );
-
-    const selection$ = events$.pipe(
-      filter(e => e.type === 'GRID/selection'),
-      map(e => e.payload as t.IGridSelectionChange),
-    );
-
-    change$.subscribe(e => {
-      // e.cancel();
-      // console.log('CHANGE', e);
-    });
-
-    change$.pipe(filter(e => e.cell.key === 'B2')).subscribe(e => {
-      console.log('B2');
-      e.cancel();
-    });
-
-    // changeSet$.subscribe(e => {});
+    const gridEvents$ = this.grid.events$.pipe(takeUntil(this.unmounted$));
+    gridEvents$.subscribe(() => this.updateState());
+    this.updateState();
   }
 
   public componentWillUnmount() {
@@ -97,49 +38,98 @@ export class Test extends React.PureComponent<ITestProps, ITestState> {
   /**
    * [Properties]
    */
-  private get Table() {
-    const { Table = HandsontableLib } = this.props;
-    return Table as Handsontable;
+  public get grid() {
+    return this.testGrid.datagrid.grid;
   }
 
-  public get grid() {
-    return this.datagrid.grid;
+  /**
+   * [Methods]
+   */
+  public updateState() {
+    const grid = this.grid;
+    const { selection, values } = grid;
+    const data = {
+      isEditing: grid.isEditing,
+      values,
+      selection,
+    };
+    this.state$.next({ data });
+    return data;
   }
 
   /**
    * [Render]
    */
   public render() {
+    const styles = {
+      base: css({
+        Flex: 'horizontal',
+        flex: 1,
+      }),
+      left: css({
+        position: 'relative',
+        width: 200,
+        padding: 10,
+        lineHeight: 1.6,
+        Flex: 'vertical-spaceBetween',
+      }),
+      leftTop: css({
+        fontSize: 13,
+      }),
+      right: css({
+        position: 'relative',
+        flex: 1,
+      }),
+      grid: css({
+        Absolute: 10,
+        border: `solid 1px ${color.format(-0.2)}`,
+      }),
+    };
+
     return (
-      <datagrid.DataGrid
-        key={'test.grid'}
-        ref={this.datagridRef}
-        values={this.state.values}
-        events$={this.events$}
-        factory={this.factory}
-        totalColumns={52}
-        totalRows={2000}
-        Handsontable={this.Table}
-        initial={{ selection: 'A1' }}
-        style={this.props.style}
-        canSelectAll={false}
-      />
+      <div {...styles.base}>
+        <div {...styles.left}>
+          <div {...styles.leftTop}>
+            {this.button('focus', () => this.grid.focus())}
+            {this.button('loadValues', () => this.grid.loadValues({ A3: 123 }))}
+            {this.button('changeValues', () => this.grid.changeValues({ A1: 'hello' }))}
+            {this.button('change values (prop)', () =>
+              this.testGrid.state$.next({ values: { A1: 'happy' } }),
+            )}
+            {this.button('select: A1', () => this.grid.select({ cell: 'A1' }))}
+            {this.button('select: A1 and range', () =>
+              this.grid.select({ cell: 'A1', ranges: ['B2:C4', 'C2:D7'] }),
+            )}
+            {this.button('select: bottom/right', () =>
+              this.grid.select({
+                cell: { row: this.grid.totalRows, column: this.grid.totalColumns },
+              }),
+            )}
+            {this.button('scrollTo: A1', () => this.grid.scrollTo({ cell: 'A1' }))}
+            {this.button('scrollTo: B5', () => this.grid.scrollTo({ cell: 'B5' }))}
+            {this.button('scrollTo: bottom/right', () =>
+              this.grid.scrollTo({
+                cell: { row: this.grid.totalRows, column: this.grid.totalColumns },
+              }),
+            )}
+          </div>
+          <ObjectView
+            name={'grid'}
+            data={this.state.data}
+            expandPaths={['$', '$.selection', '$.selection.ranges']}
+          />
+        </div>
+        <div {...styles.right}>
+          <TestGridView ref={this.testGridRef} style={styles.grid} />
+        </div>
+      </div>
     );
   }
 
-  private factory: t.GridFactory = req => {
-    switch (req.type) {
-      case 'EDITOR':
-        return <DebugEditor />;
-
-      case 'CELL':
-        const value = typeof req.value === 'object' ? JSON.stringify(req.value) : req.value;
-        // return <div>{value}</div>;
-        return value;
-
-      default:
-        console.log(`Factory type '${req.type}' not supported by test.`);
-        return null;
-    }
+  /**
+   * [Handlers]
+   */
+  private button = (label: string, handler: () => void) => {
+    return <Button label={label} onClick={handler} block={true} />;
   };
 }
