+++ code/ui.datagrid/test/components/Test.tsx
@@ -1,36 +1,37 @@
+import { CommandPrompt } from '@platform/ui.cli';
 import * as React from 'react';
 import { Subject } from 'rxjs';
-import { filter, takeUntil } from 'rxjs/operators';
+import { takeUntil } from 'rxjs/operators';
 
-import { color, css, Button, ObjectView, COLORS } from '../common';
-import { Test as TestGrid } from './Test.Grid';
-import { CommandPrompt } from '@platform/ui.cli';
 import { init as initCommandLine } from '../cli';
+import { color, COLORS, css, t } from '../common';
+import { TestCellEditor } from './Test.CellEditor';
+import { TestGrid } from './Test.Grid';
 
-export type ITestProps = {};
-export type ITestState = {
-  data?: any;
+const KEY = {
+  VIEW: 'ui.datagrid/view',
 };
 
-export class Test extends React.PureComponent<ITestProps, ITestState> {
-  public state: ITestState = {};
+export type ITestProps = {};
+
+export class Test extends React.PureComponent<ITestProps, t.ITestState> {
+  public state: t.ITestState = {};
   private unmounted$ = new Subject();
-  private state$ = new Subject<Partial<ITestState>>();
-
-  private testGrid!: TestGrid;
-  private testGridRef = (ref: TestGrid) => (this.testGrid = ref);
-
-  private cli = initCommandLine({});
+  private state$ = new Subject<Partial<t.ITestState>>();
+  private cli = initCommandLine({ state$: this.state$ });
 
   /**
    * [Lifecycle]
    */
 
-  public componentDidMount() {
-    this.state$.pipe(takeUntil(this.unmounted$)).subscribe(e => this.setState(e));
-    const events$ = this.grid.events$.pipe(takeUntil(this.unmounted$));
-    events$.pipe().subscribe(() => this.updateState());
-    this.updateState();
+  public componentWillMount() {
+    const state$ = this.state$.pipe(takeUntil(this.unmounted$));
+    state$.subscribe(e => this.setState(e));
+
+    // Save and resume the current view using local-storage.
+    state$.subscribe(() => localStorage.setItem(KEY.VIEW, this.view));
+    const view = (localStorage.getItem(KEY.VIEW) as t.ITestState['view']) || this.view;
+    this.state$.next({ view });
   }
 
   public componentWillUnmount() {
@@ -40,23 +41,9 @@ export class Test extends React.PureComponent<ITestProps, ITestState> {
   /**
    * [Properties]
    */
-  public get grid() {
-    return this.testGrid.datagrid.grid;
-  }
 
-  /**
-   * [Methods]
-   */
-  public updateState() {
-    const grid = this.grid;
-    const { selection, values } = grid;
-    const data = {
-      isEditing: grid.isEditing,
-      values,
-      selection,
-    };
-    this.state$.next({ data });
-    return data;
+  public get view() {
+    return this.state.view || 'grid';
   }
 
   /**
@@ -71,67 +58,23 @@ export class Test extends React.PureComponent<ITestProps, ITestState> {
       }),
       main: css({
         position: 'relative',
-        Flex: 'horizontal',
+        display: 'flex',
         flex: 1,
-      }),
-      left: css({
-        position: 'relative',
-        width: 200,
-        padding: 10,
-        lineHeight: 1.6,
-        Flex: 'vertical-spaceBetween',
-      }),
-      leftTop: css({
-        fontSize: 13,
-      }),
-      right: css({
-        position: 'relative',
-        flex: 1,
-      }),
-      grid: css({
-        Absolute: 10,
-        border: `solid 1px ${color.format(-0.2)}`,
       }),
       footer: css({
         backgroundColor: COLORS.DARK,
       }),
     };
+
+    const view = this.view;
+    const elGrid = view === 'grid' && <TestGrid />;
+    const elCellEditor = view === 'editor' && <TestCellEditor />;
+
     return (
       <div {...styles.base}>
         <div {...styles.main}>
-          <div {...styles.left}>
-            <div {...styles.leftTop}>
-              {this.button('focus', () => this.grid.focus())}
-              {this.button('loadValues', () => this.grid.loadValues({ A3: 123 }))}
-              {this.button('changeValues', () => this.grid.changeValues({ A1: 'hello' }))}
-              {this.button('change values (prop)', () =>
-                this.testGrid.state$.next({ values: { A1: 'happy' } }),
-              )}
-              {this.button('select: A1', () => this.grid.select({ cell: 'A1' }))}
-              {this.button('select: A1 and range', () =>
-                this.grid.select({ cell: 'A1', ranges: ['B2:C4', 'C2:D7'] }),
-              )}
-              {this.button('select: bottom/right', () =>
-                this.grid.select({
-                  cell: { row: this.grid.totalRows, column: this.grid.totalColumns },
-                }),
-              )}
-              {this.button('scrollTo: A1', () => this.grid.scrollTo({ cell: 'A1' }))}
-              {this.button('scrollTo: B5', () => this.grid.scrollTo({ cell: 'B5' }))}
-              {this.button('scrollTo: bottom/right', () =>
-                this.grid.scrollTo({
-                  cell: { row: this.grid.totalRows, column: this.grid.totalColumns },
-                }),
-              )}
-            </div>
-            <ObjectView
-              data={this.state.data}
-              expandPaths={['$', '$.selection', '$.selection.ranges']}
-            />
-          </div>
-          <div {...styles.right}>
-            <TestGrid ref={this.testGridRef} style={styles.grid} />
-          </div>
+          {elGrid}
+          {elCellEditor}
         </div>
         <div {...styles.footer}>
           <CommandPrompt cli={this.cli} theme={'DARK'} />
@@ -139,19 +82,4 @@ export class Test extends React.PureComponent<ITestProps, ITestState> {
       </div>
     );
   }
-
-  /**
-   * [Handlers]
-   */
-  private button = (label: string, handler: () => void) => {
-    return (
-      <div>
-        <Button label={label} onClick={handler} />
-      </div>
-    );
-  };
-
-  private tmp = () => {
-    this.grid.select({ cell: 'A1' });
-  };
 }
