+++ code/ui.datagrid/test/types.ts
@@ -1,0 +1,12 @@
+import { Subject } from 'rxjs';
+
+export * from '../src/types';
+export * from '@platform/ui.cli/lib/types';
+
+export type ITestCommandProps = {
+  state$: Subject<Partial<ITestState>>;
+};
+
+export type ITestState = {
+  view?: 'grid' | 'editor';
+};
